+++ mne/viz/tests/test_epochs.py
@@ -144,6 +144,7 @@ def test_plot_epochs():
         _fake_click(fig, data_ax, [x, y], xform='data')  # unmark a bad epoch
         _fake_click(fig, data_ax, [0.5, 0.999])  # click elsewhere in 1st axes
         _fake_click(fig, data_ax, [-0.1, 0.9])  # click on y-label
+        _fake_click(fig, data_ax, [-0.1, 0.9], button=3)
         _fake_click(fig, fig.get_axes()[2], [0.5, 0.5])  # change epochs
         _fake_click(fig, fig.get_axes()[3], [0.5, 0.5])  # change channels
         fig.canvas.close_event()  # closing and epoch dropping