+++ doc/whats_new.rst
@@ -39,8 +39,6 @@ Changelog
 
     - Add new method :class:`mne.preprocessing.Xdawn` for denoising and decoding of ERP/ERF by `Alexandre Barachant`_
 
-    - Add ``plot_topo`` method to ``Evoked`` object by `Jaakko Leppakangas`_
-
     - Add support for plotting patterns/filters in :class:`mne.decoding.csp.CSP` and :class:`mne.decoding.base.LinearModel` by `Romain Trachel`_
 
     - Add new object :class:`mne.decoding.base.LinearModel` for decoding M/EEG data and interpreting coefficients of linear models with patterns attribute by `Romain Trachel`_ and `Alex Gramfort`_
@@ -57,6 +55,8 @@ Changelog
 
     - Add fetcher :mod:`mne.datasets.brainstorm` for datasets used by Brainstorm in their tutorials by `Mainak Jas`_
 
+    - Add interactive plotting of single trials by right clicking on channel name in epochs browser by `Jaakko Leppakangas`_
+
 BUG
 ~~~
 
@@ -69,6 +69,8 @@ BUG
     - Fix picking channels from forward operator could return a channel ordering different from ``info['chs']`` by `Chris Bailey`_
 
     - Fix dropping of events after downsampling stim channels by `Marijn van Vliet`
+
+    - Fix scaling in :func:``mne.viz.utils._setup_vmin_vmax`` by `Jaakko Leppakangas`_
 
 API
 ~~~